# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ResumeUpload {
  file: Blob;
}

export type ResumeStatus = typeof ResumeStatus[keyof typeof ResumeStatus];


export const ResumeStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface ExtractedSkill {
  name: string;
  confidence: number;
  category: string;
  source?: string;
}

export interface Resume {
  id: number;
  filename: string;
  originalName: string;
  status: ResumeStatus;
  createdAt: string;
  /** @nullable */
  extractedText?: string | null;
  skills?: ExtractedSkill[];
  education?: string[];
  experience?: string[];
  certifications?: string[];
  /** @nullable */
  yearsOfExperience?: number | null;
}

export interface AnalysisRequest {
  resumeId: number;
  jobDescription: string;
  jobTitle?: string;
}

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  completed: 'completed',
  error: 'error',
} as const;

export interface Analysis {
  id: number;
  resumeId: number;
  status: AnalysisStatus;
  createdAt: string;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  atsScore?: number | null;
  /** @nullable */
  missingSkillsCount?: number | null;
}

export type AnalysisDetailStatus = typeof AnalysisDetailStatus[keyof typeof AnalysisDetailStatus];


export const AnalysisDetailStatus = {
  pending: 'pending',
  completed: 'completed',
  error: 'error',
} as const;

export type SkillGapItemPriority = typeof SkillGapItemPriority[keyof typeof SkillGapItemPriority];


export const SkillGapItemPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface SkillGapItem {
  name: string;
  category: string;
  priority: SkillGapItemPriority;
  confidence?: number;
  /** @nullable */
  demandScore?: number | null;
}

export interface AnalysisDetail {
  id: number;
  resumeId: number;
  status: AnalysisDetailStatus;
  createdAt: string;
  /** @nullable */
  jobTitle?: string | null;
  jobDescription?: string;
  matchScore: number;
  atsScore: number;
  missingSkills?: SkillGapItem[];
  strongSkills?: SkillGapItem[];
  improvementAreas?: string[];
  atsRecommendations?: string[];
  resume?: Resume;
}

export type LearningResourceType = typeof LearningResourceType[keyof typeof LearningResourceType];


export const LearningResourceType = {
  course: 'course',
  article: 'article',
  documentation: 'documentation',
  video: 'video',
  practice: 'practice',
} as const;

export interface LearningResource {
  title: string;
  type: LearningResourceType;
  url: string;
  provider?: string;
  isFree?: boolean;
}

export type RoadmapPhasePriority = typeof RoadmapPhasePriority[keyof typeof RoadmapPhasePriority];


export const RoadmapPhasePriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface RoadmapPhase {
  skill: string;
  description: string;
  resources: LearningResource[];
  priority: RoadmapPhasePriority;
  estimatedHours: number;
  category?: string;
}

export interface LearningRoadmap {
  analysisId: number;
  thirtyDayPlan: RoadmapPhase[];
  sixtyDayPlan: RoadmapPhase[];
  ninetyDayPlan: RoadmapPhase[];
}

export interface DataSource {
  name: string;
  description: string;
  lastUpdated: string;
  confidenceScore: number;
  category?: string;
}

export type SkillTrendTrendDirection = typeof SkillTrendTrendDirection[keyof typeof SkillTrendTrendDirection];


export const SkillTrendTrendDirection = {
  rising: 'rising',
  stable: 'stable',
  declining: 'declining',
} as const;

export interface SkillTrend {
  name: string;
  demandScore: number;
  trendDirection: SkillTrendTrendDirection;
  category: string;
  jobCount: number;
  growthRate?: number;
}

export interface CategoryDemand {
  category: string;
  demandScore: number;
  skillCount: number;
  topSkills?: string[];
}

export interface MarketTrendsResponse {
  trendingSkills: SkillTrend[];
  emergingSkills: SkillTrend[];
  topCategories: CategoryDemand[];
  lastUpdated: string;
}

export interface SkillFrequency {
  skill: string;
  count: number;
  category: string;
}

export interface DashboardStats {
  totalResumes: number;
  totalAnalyses: number;
  avgAtsScore: number;
  avgMatchScore: number;
  topMissingSkills: SkillFrequency[];
  recentAnalyses: Analysis[];
}