# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from './analysisStatus';

export interface Analysis {
  id: number;
  resumeId: number;
  status: AnalysisStatus;
  createdAt: Date;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  atsScore?: number | null;
  /** @nullable */
  missingSkillsCount?: number | null;
}