# lib/api-zod/src/generated/types/learningResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningResourceType } from './learningResourceType';

export interface LearningResource {
  title: string;
  type: LearningResourceType;
  url: string;
  provider?: string;
  isFree?: boolean;
}