# lib/api-zod/src/generated/types/learningRoadmap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapPhase } from './roadmapPhase';

export interface LearningRoadmap {
  analysisId: number;
  thirtyDayPlan: RoadmapPhase[];
  sixtyDayPlan: RoadmapPhase[];
  ninetyDayPlan: RoadmapPhase[];
}