# lib/api-zod/src/generated/types/marketTrendsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryDemand } from './categoryDemand';
import type { SkillTrend } from './skillTrend';

export interface MarketTrendsResponse {
  trendingSkills: SkillTrend[];
  emergingSkills: SkillTrend[];
  topCategories: CategoryDemand[];
  lastUpdated: string;
}