# lib/api-zod/src/generated/types/roadmapPhase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningResource } from './learningResource';
import type { RoadmapPhasePriority } from './roadmapPhasePriority';

export interface RoadmapPhase {
  skill: string;
  description: string;
  resources: LearningResource[];
  priority: RoadmapPhasePriority;
  estimatedHours: number;
  category?: string;
}