# lib/api-zod/src/generated/types/skillFrequency.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillFrequency {
  skill: string;
  count: number;
  category: string;
}