# lib/api-zod/src/generated/types/skillGapItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillGapItemPriority } from './skillGapItemPriority';

export interface SkillGapItem {
  name: string;
  category: string;
  priority: SkillGapItemPriority;
  confidence?: number;
  /** @nullable */
  demandScore?: number | null;
}